feat: implementando caso de uso de busca de parágrafos similares à query passada, desenvolvendo busca híbrida, e rota para API com FastAPI.
Este commit tem como objetivo desenvolver o caso de uso para busca de parágrafos similares à query passada no banco de dados vetorial Qdrant, através de busca híbrida. Este commit também tem como objetivo começar a desenvolver a API com FastAPI para captura de parágrafos similares à query enviada como requsição.
Este commit surge da necessidade de versionar mais uma feature implementada, e funcional deste projeto.

+++ src/infra/vector_db/qdrant/qdrant_vector_db_repository.py
@@ -1,6 +1,11 @@
-from typing import List, Dict
-from qdrant_client.http.models import PointStruct, Prefetch
+from typing import List
+from qdrant_client.http.models import PointStruct
 from src.validators.models.SearchOutput import SearchOutput
+from src.validators.models.QueryEmbedding import QueryEmbeddingBase
+from src.errors.types.collection_not_found_error import CollectionNotFoundError
+from src.errors.types.points_searcher_not_found_error import PointsSearcherNotFoundError
+from src.errors.types.collection_creator_not_found_error import CollectionCreatorNotFoundError
+from src.infra.interfaces.qdrant_points_searcher_interface import QdrantPointsSearcherInterface
 from src.infra.interfaces.qdrant_collection_creator_interface import QdrantCollectionCreatorInterface
 from src.infra.vector_db.qdrant.settings.qdrant_vector_db_connection_handler import QdrantVectorDBConnectionHandler
 
@@ -9,14 +14,20 @@ logger = setup_logger(name="QdrantVectorDBRepository")
 
 
 class QdrantVectorDBRepository:
-    def __init__(self, collection_name: str, collection_creator: QdrantCollectionCreatorInterface):
+    def __init__(self, collection_name: str, collection_creator: QdrantCollectionCreatorInterface | None = None,
+                 points_searcher: QdrantPointsSearcherInterface | None = None) -> None:
         self.collection_name = collection_name
         self.collection_creator = collection_creator
+        self.points_searcher = points_searcher
 
     def create_collection(self) -> None:
+        if self.collection_creator is None:
+            raise CollectionCreatorNotFoundError(
+                f"O criador de coleção não foi passado. O criador de coleção não pode ser None.")
+
         if self.__collection_already_exist() and self.__collection_already_populated():
             logger.info(
-                f"A coleção '{self.collection_name}' já existe existe e já possui dados. "
+                f"A coleção '{self.collection_name}' já existe existe e já possui dados."
                 f"Ela será excluída e criada novamente.")
             self.__delete_collection()
             self.collection_creator.create(
@@ -25,82 +36,71 @@ class QdrantVectorDBRepository:
             self.collection_creator.create(
                 collection_name=self.collection_name)
 
-    def upsert_data(self, ingestion_points: List[PointStruct]):
-        with QdrantVectorDBConnectionHandler() as qdrant:
-            try:
-                logger.info(
-                    f"  Enviando {len(ingestion_points)} estruturas de pontos de ingestão para '{self.collection_name}'...")
+    def upsert_points(self, ingestion_points: List[PointStruct]) -> None:
+        if not self.__collection_already_exist():
+            raise CollectionNotFoundError(
+                f"A coleção '{self.collection_name}' não existe para que estruras de pontos possam ser armazenadas.")
 
+        with QdrantVectorDBConnectionHandler() as qdrant:
+            logger.info(
+                f"  Enviando {len(ingestion_points)} estruturas de pontos de ingestão para '{self.collection_name}'...")
+            try:
                 qdrant.client.upsert(
                     collection_name=self.collection_name, points=ingestion_points)
 
                 logger.info(
                     "  As estruturas de pontos de ingestão foram enviados com sucesso!")
+            except Exception as exception:
+                logger.exception(f"Exceção ao inserir as estruturas de pontos de ingestão na coleção '{self.collection_name}'.\n"
+                                 f"Exception: {exception}")
+                raise
 
+            try:
                 collection_info = qdrant.client.get_collection(
                     self.collection_name)
                 logger.info(
                     f"A coleção '{self.collection_name}' agora possui {collection_info.points_count} parágrafos.")
-
-            except Exception as error:
-                logger.error(f"Erro ao inserir as estruturas de pontos de ingestão na coleção '{self.collection_name}'.\n"
-                             f"Error Message: {error}")
+            except Exception as exception:
+                logger.exception(f"Exceção ao obter informações sobre a coleção '{self.collection_name}'.\n"
+                                 f"Exception: {exception}")
                 raise
 
-    def hybrid_data_search(self, collection_name: str, prefetch_limit: int, embeddings: Dict, top_k: int = 5):
-        with QdrantVectorDBConnectionHandler() as qdrant:
-            try:
-                search_result = qdrant.client.query_points(
-                    collection_name=collection_name,
-                    prefetch=[
-                        Prefetch(
-                            query=embeddings['dense'],
-                            using="dense",
-                            limit=prefetch_limit,
-                        ),
-                        Prefetch(
-                            query=embeddings['sparse'].model_dump(),
-                            using="sparse",
-                            limit=prefetch_limit,
-                        ),
-                    ],
-                    query=embeddings['late'],
-                    using="colbertv2.0",
-                    with_payload=True,
-                    limit=top_k,
-                )
+    def search_points(self, embedding: QueryEmbeddingBase) -> List[SearchOutput]:
+        if self.points_searcher is None:
+            raise PointsSearcherNotFoundError(
+                f"O buscador de pontos de vetoriais não foi passado. O buscador não pode ser None.")
 
-                search_outputs_list = [SearchOutput(
-                    text=point.payload['text'],
-                    metadata=point.payload['metadata']
-                ) for point in search_result.points if point.payload != None]
+        if not self.__collection_already_populated():
+            raise CollectionNotFoundError(
+                f"A coleção '{self.collection_name}' não existe! Assim, não há com realizar a busca de pontos vetoriais.")
 
-                return [output.model_dump() for output in search_outputs_list]
+        search_outputs: List[SearchOutput] = self.points_searcher.search(
+            collection_name=self.collection_name,
+            embedding=embedding)
 
-            except Exception as error:
-                logger.error(f"Erro ao realizar busca híbrida na coleção {collection_name}.\n"
-                             f"Error Message: {error}")
-                raise
+        return search_outputs
 
     def __collection_already_exist(self) -> bool:
         with QdrantVectorDBConnectionHandler() as qdrant:
             try:
                 return qdrant.client.collection_exists(collection_name=self.collection_name)
-            except Exception as error:
-                logger.info(f"Ocorreu uma falha ao verificar se a coleção '{self.collection_name}' no Qdrant existe.\n"
-                            f"Error Message: {error}.")
+            except Exception as exception:
+                logger.exception(f"Ocorreu uma falha ao verificar se a coleção '{self.collection_name}' no Qdrant existe.\n"
+                                 f"Exception: {exception}.")
                 raise
 
     def __collection_already_populated(self) -> bool:
-        if self.__collection_already_exist():
-            with QdrantVectorDBConnectionHandler() as qdrant:
-                try:
-                    return int(qdrant.client.count(collection_name=self.collection_name).count) > 0
-                except Exception as error:
-                    logger.info(f"Ocorreu uma falha ao verificar se a coleção '{self.collection_name}' no Qdrant já está populada.\n"
-                                f"Error Message: {error}.")
-                    raise
-        return False
+        if not self.__collection_already_exist():
+            raise CollectionNotFoundError(
+                f"A coleção '{self.collection_name}' não existe! Com isso, não dá para checar se a mesma está populada.")
+
+        with QdrantVectorDBConnectionHandler() as qdrant:
+            try:
+                return int(qdrant.client.count(collection_name=self.collection_name).count) > 0
+            except Exception as exception:
+                logger.exception(f"Ocorreu uma falha ao verificar se a coleção '{self.collection_name}' no Qdrant já está populada.\n"
+                                 f"Exception: {exception}.")
+                raise
 
     def __delete_collection(self) -> None:
         with QdrantVectorDBConnectionHandler() as qdrant:
@@ -109,7 +109,7 @@ class QdrantVectorDBRepository:
                     collection_name=self.collection_name)
                 logger.info(
                     f"A coleção '{self.collection_name}' foi excluída com sucesso.")
-            except Exception as error:
-                logger.error(f"Ocorreu uma falha ao excluir a coleção '{self.collection_name}' do Qdrant.\n"
-                             f"Error Message: {error}.")
+            except Exception as exception:
+                logger.exception(f"Ocorreu uma falha ao excluir a coleção '{self.collection_name}' do Qdrant.\n"
+                                 f"Exception: {exception}.")
                 raise
